% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preprints}
\alias{preprints}
\title{PsyArXiv reprint metadata from the Open Science Framework.}
\format{
A data frame with 15 variables:
\describe{
\item{\code{id}}{Preprint identifier (chr).}
\item{\code{title}}{Preprint title (chr).}
\item{\code{description}}{Preprint description or abstract (chr).}
\item{\code{date_created}}{The time at which the preprint was created, as an iso8601 formatted timestamp. (dttm).}
\item{\code{date_modified}}{The time at which the preprint was last modified, as an iso8601 formatted timestamp. (dttm).}
\item{\code{date_published}}{The time at which the preprint was published, as an iso8601 formatted timestamp. (dttm).}
\item{\code{original_publication_date}}{When published elsewhere (dttm).}
\item{\code{publication_doi}}{DOI of document published elsewhere (chr).}
\item{\code{preprint_doi}}{Preprint's DOI with full URL (chr).}
\item{\code{download_url}}{URL to download preprint document (chr).}
\item{\code{license}}{License (chr).}
\item{\code{is_published}}{Is preprint published on PsyArXiv.}
\item{\code{reviews_state}}{Is preprint reviewed/accepted at PsyArXiv.}
\item{\code{version}}{Preprint version (int).}
\item{\code{is_latest_version}}{Is this the preprint's latest version (lgl).}
\item{\code{has_coi}}{A value indicating whether the preprint has a conflict of interest statement.}
\item{\code{conflict_of_interest_statement}}{The text of the conflict of interest statement for the preprint.}
\item{\code{has_data_links}}{A value indicating whether the preprint has a links to data.}
\item{\code{why_no_data}}{A value for a field explaining why preregistration data was not included.}
\item{\code{data_links}}{Any links to a preprints supplemental data go here.}
\item{\code{has_prereg_links}}{A value indicating whether the preprint has any preregistration links.}
\item{\code{why_no_prereg}}{A value for a field explaining why preregistration links weren't included.}
\item{\code{prereg_links}}{Any links to a preprints preregistrations go here.}
\item{\code{prereg_link_info}}{The text of the any explanation of preprint preregistration link infomation.}
\item{\code{contributors}}{Contributor metadata (chr/JSON).}
\item{\code{subjects}}{Subject metadata (chr/JSON).}
\item{\code{tags}}{Tags (chr/JSON).}
}
}
\usage{
preprints
}
\value{
a \link[tibble:tibble-package]{tibble}

For further details, see \url{https://developer.osf.io/#tag/Preprints}.
}
\description{
Preprint metadata from the Open Science Framework. All data licensed to preprint authors as indicated.
}
\keyword{datasets}
