% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psyarxivdbr-package.R
\docType{package}
\name{psyarxivdbr-package}
\alias{psyarxivdbr}
\alias{psyarxivdbr-package}
\title{psyarxivdbr: PsyArXiv preprints data}
\description{
Convenient access to PsyArXiv preprint data from the Open Science Framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mvuorre/psyarxivdbr}
  \item Report bugs at \url{https://github.com/mvuorre/psyarxivdbr/issues}
}

}
\author{
\strong{Maintainer}: Matti Vuorre \email{mjvuorre@uvt.nl} (\href{https://orcid.org/0000-0001-5052-066X}{ORCID})

}
\keyword{internal}
