% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psyarxivr-package.R
\docType{package}
\name{psyarxivr-package}
\alias{psyarxivr}
\alias{psyarxivr-package}
\title{psyarxivr: PsyArXiv preprints data}
\description{
Convenient access to PsyArXiv preprint data from the Open Science Framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mvuorre/psyarxivr}
  \item Report bugs at \url{https://github.com/mvuorre/psyarxivr/issues}
}

}
\author{
\strong{Maintainer}: Matti Vuorre \email{mjvuorre@uvt.nl} (\href{https://orcid.org/0000-0001-5052-066X}{ORCID})

}
\keyword{internal}
